requirements.txt
beautifulsoup4>=4.9.1
python-dotenv>=0.14.0
python-slugify>=4.0.1
requests>=2.23.0